+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix unstable comments in between Try/Catch/Finally blocks.
 - Fix unstable comments in WhereCompoundExpr.
 - Fix unstable comments for NestedExpr.
+- Fix unstable formatting for Method Declaration with no body.
 
 ## 1.0.0-beta.2
 - Fix long static method calls producing undeployable code ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/69)).